refactor: CardNumber가 view라고 생각하여 view로 패키지 이동 및 수정
