+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,5 +1,7 @@
 package blackjack.domain
 
+import blackjack.view.CardSuite
+
 data class Card(val suite: CardSuite, val number: CardNumber) {
 
     companion object {
