+++ src/main/kotlin/blackjack/domain/CardNumber.kt
@@ -1,17 +1,17 @@
 package blackjack.domain
 
-enum class CardNumber(val value: Int, val displayName: String) {
-    TWO(2, "2"),
-    THREE(3, "3"),
-    FOUR(4, "4"),
-    FIVE(5, "5"),
-    SIX(6, "6"),
-    SEVEN(7, "7"),
-    EIGHT(8, "8"),
-    NINE(9, "9"),
-    TEN(10, "10"),
-    JACK(10, "J"),
-    QUEEN(10, "Q"),
-    KING(10, "K"),
-    ACE(1, "A")
+enum class CardNumber(val value: Int) {
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5),
+    SIX(6),
+    SEVEN(7),
+    EIGHT(8),
+    NINE(9),
+    TEN(10),
+    JACK(10),
+    QUEEN(10),
+    KING(10),
+    ACE(1)
 }
