+++ src/test/kotlin/blackjack/domain/CardSuiteTests.kt
@@ -1,5 +1,6 @@
 package blackjack.domain
 
+import blackjack.view.CardSuite
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
