+++ src/test/kotlin/blackjack/domain/CardTest.kt
@@ -1,5 +1,6 @@
 package blackjack.domain
 
+import blackjack.view.CardSuite
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
