+++ src/test/kotlin/blackjack/domain/GameTest.kt
@@ -1,6 +1,7 @@
 package blackjack.domain
 
 import blackjack.fixture.PlayFixture.TEST_NAME
+import blackjack.view.CardSuite
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
